requirements.txt
streamlit>=1.30.0
ipykernel>=6
pandas>=1.5
numpy>=1.21
matplotlib>=3.5
plotly>=5
seaborn>=0.12
scikit-learn>=1.0
pymongo>=4
from_root
dill>=0.3
certifi>=2021.10
PyYAML>=6
boto3>=1.20
botocore>=1.23
fastapi>=0.70
python-multipart>=0.0.5
uvicorn>=0.17
jinja2>=3
imblearn>=0.0
